テストファイルの削除・ファイルのリネーム
