+++ src/main/Cpu.class.ts
@@ -1,8 +1,6 @@
-import { Field } from "./Field.class";
-// import { Mino } from "./Mino.class";
-// import { Wetris } from "./Wetris.class";
-import { nWetrisCore } from "./nWetrisCore";
-import { nWetrisSender } from "./nWetrisSender";
+import { Field } from "./Field";
+import { WetrisCore } from "./WetrisCore";
+import { WetrisSender } from "./WetrisSender";
 
 import { MINO_IDX, DRAW_FIELD_TOP, ARR } from "./constant";
 
@@ -28,12 +26,12 @@ type FieldInfo = {
 type FieldScore = { fieldData: FieldData; score: number };
 
 export class Cpu {
-    mainWetris: nWetrisSender;
-    trialWetris: nWetrisCore;
+    mainWetris: WetrisSender;
+    trialWetris: WetrisCore;
 
-    constructor(wetris: nWetrisSender) {
+    constructor(wetris: WetrisSender) {
         this.mainWetris = wetris;
-        this.trialWetris = new nWetrisCore();
+        this.trialWetris = new WetrisCore();
         this.trialWetris.isMainloopActive = false;
         this.main();
     }
@@ -54,7 +52,7 @@ export class Cpu {
         }
     }
 
-    async moveMinoToMatchField(wetris: nWetrisSender, fieldData: FieldData) {
+    async moveMinoToMatchField(wetris: WetrisSender, fieldData: FieldData) {
         while (wetris.currentMino.angle % 4 !== fieldData.angle % 4) {
             wetris.rotateRight();
             await wetris.sleep(ARR);
@@ -90,7 +88,7 @@ export class Cpu {
         for (let angle = 0; angle < 4; angle++) {
             // 左から順に、移動可能な全てのx座標における一番下に接地した場合を調べる
             for (let movement = 0; ; movement++) {
-                this.trialWetris = new nWetrisCore();
+                this.trialWetris = new WetrisCore();
                 this.trialWetris.isMainloopActive = false;
                 this.trialWetris.currentMino.idxMino = idxMino;
                 for (let i = 0; i < angle; i++) {
